also move the syncstate package to ./internal/ingest
