+++ internal/ingest/syncstate/encoding.go
