+++ internal/ingest/syncstate/encoding_test.go
